+++ src/main/java/shadows/apotheosis/Apotheosis.java
@@ -30,8 +30,13 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.simple.SimpleChannel;
+import net.minecraftforge.registries.ForgeRegistry;
+import net.minecraftforge.registries.NewRegistryEvent;
+import net.minecraftforge.registries.RegistryBuilder;
 import shadows.apotheosis.advancements.AdvancementTriggers;
 import shadows.apotheosis.adventure.AdventureModule;
+import shadows.apotheosis.adventure.affix.Affix;
+import shadows.apotheosis.adventure.affix.AffixHelper;
 import shadows.apotheosis.adventure.client.BossSpawnMessage;
 import shadows.apotheosis.compat.PatchouliCompat;
 import shadows.apotheosis.ench.EnchModule;
@@ -112,6 +117,7 @@ public class Apotheosis {
 		if (config.hasChanged()) config.save();
 		bus.post(new ApotheosisConstruction());
 		bus.addListener(this::init);
+		bus.addListener(this::registry);
 		MinecraftForge.EVENT_BUS.addListener(this::reloads);
 		MinecraftForge.EVENT_BUS.addListener(this::trackCooldown);
 		MinecraftForge.EVENT_BUS.addListener(this::cmds);
@@ -127,6 +133,15 @@ public class Apotheosis {
 		e.enqueueWork(AdvancementTriggers::init);
 		CraftingHelper.register(new ModuleCondition.Serializer());
 		CraftingHelper.register(new ResourceLocation(MODID, "enchantment"), EnchantmentIngredient.Serializer.INSTANCE);
+	}
+
+	@SubscribeEvent
+	public void registry(NewRegistryEvent e) {
+		RegistryBuilder<Affix> build = new RegistryBuilder<>();
+		build.setName(new ResourceLocation(Apotheosis.MODID, "affixes"));
+		build.setType(Affix.class);
+		build.disableSaving().onBake(AffixHelper::recomputeMaps);
+		e.create(build, r -> Affix.REGISTRY = (ForgeRegistry<Affix>) r);
 	}
 
 	@SubscribeEvent