+++ src/main/java/shadows/apotheosis/adventure/AdventureConfig.java
@@ -41,9 +41,10 @@ public class AdventureConfig {
 	// Affix
 	public static float randomAffixItem = 0.07F;
 	public static float gemDropChance = 0.04F;
-	public static float affixChestChance = 0.75F;
-	public static float gemChestChance = 0.55F;
-	public static int[] rarityThresholds = new int[] { 400, 700, 850, 940, 995 };
+	public static float gemBossBonus = 0.33F;
+	public static float affixChestChance = 0.55F;
+	public static float gemChestChance = 0.35F;
+	public static int[] rarityThresholds = new int[] { 400, 720, 870, 960, 995 };
 	public static boolean disableQuarkOnAffixItems = true;
 	public static Supplier<Item> torchItem = () -> Items.TORCH;
 
@@ -72,6 +73,7 @@ public class AdventureConfig {
 
 		randomAffixItem = c.getFloat("Random Affix Chance", "affixes", randomAffixItem, 0, 1, "The chance that a naturally spawned mob will be granted an affix item. 0 = 0%, 1 = 100%");
 		gemDropChance = c.getFloat("Gem Drop Chance", "affixes", gemDropChance, 0, 1, "The chance that a mob will drop a gem. 0 = 0%, 1 = 100%");
+		gemBossBonus = c.getFloat("Gem Boss Bonus", "affixes", gemBossBonus, 0, 1, "The flat bonus chance that bosses have to drop a gem, added to Gem Drop Chance. 0 = 0%, 1 = 100%");
 
 		affixChestChance = c.getFloat("Affix Chest Chance", "affixes", gemChestChance, 0, 1, "The chance that an affix item will be added to a loot chest. 0 = 0%, 1 = 100%");
 		gemChestChance = c.getFloat("Gem Chest Chance", "affixes", gemChestChance, 0, 1, "The chance that a gem will be added to a loot chest. 0 = 0%, 1 = 100%");