+++ src/main/java/shadows/apotheosis/adventure/boss/BossDungeonFeature2.java
@@ -61,7 +61,7 @@ public class BossDungeonFeature2 extends Feature<NoneFeatureConfiguration> {
 					BlockState state = world.getBlockState(blockpos);
 					Material material = state.getMaterial();
 					boolean flag = material.isSolid();
-					 //Exit if the floor is not fully solid.
+					//Exit if the floor is not fully solid.
 
 					if ((y == floor && !flag) || (y == roof && !flag)) { return false; } //Exit if the roof is not fully solid.
 					if (y == roof + 1 && Math.abs(x) < xRadius && Math.abs(z) < zRadius && !flag) { return false; } //Exit if the roof is not fully solid.
@@ -94,7 +94,7 @@ public class BossDungeonFeature2 extends Feature<NoneFeatureConfiguration> {
 			world.setBlock(pos, Apoth.Blocks.BOSS_SPAWNER.defaultBlockState(), 2);
 			BlockEntity tileentity = world.getBlockEntity(pos);
 			if (tileentity instanceof BossSpawnerTile) {
-				((BossSpawnerTile) tileentity).setBossItem(BossItemManager.INSTANCE.getRandomItem(rand));
+				((BossSpawnerTile) tileentity).setBossItem(BossItemManager.INSTANCE.getRandomItem(rand, world));
 			} else {
 				AdventureModule.LOGGER.error("Failed to fetch boss spawner entity at ({}, {}, {})", pos.getX(), pos.getY(), pos.getZ());
 			}