+++ src/main/java/shadows/apotheosis/adventure/loot/LootCategory.java
@@ -19,6 +19,7 @@ import net.minecraft.world.item.TridentItem;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.common.ToolActions;
 import net.minecraftforge.server.ServerLifecycleHooks;
+import shadows.apotheosis.adventure.AdventureConfig;
 
 public enum LootCategory {
 	NONE(Predicates.alwaysFalse(), s -> new EquipmentSlot[0]),
@@ -50,6 +51,8 @@ public enum LootCategory {
 	static final LootCategory[] VALUES = values();
 
 	public static LootCategory forItem(ItemStack item) {
+		LootCategory override = AdventureConfig.TYPE_OVERRIDES.get(item.getItem().getRegistryName());
+		if (override != null) return override;
 		for (LootCategory c : VALUES) {
 			if (c.isValid(item)) return c;
 		}